refactor: move db configuration files to src/lib/database
